+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/referee_game_files.js
@@ -56,7 +56,9 @@ var buildPlayerFiles = function (params, gwoAI, gwoSpecs) {
   var inventory = params.inventory;
   var titans = params.titans;
 
-  var playerIsCluster = inventory.getTag("global", "playerFaction") === 4;
+  var playerIsCluster = gwoAI.isCluster({
+    faction: inventory.getTag("global", "playerFaction"),
+  });
   var hostSubcommanderPath = gwoAI.getAIPathDestination("subcommander");
   var playerFilesClassic;
   var playerFilesX1;